backend/ai-server/requirements.txt
Flask==2.2.2
Flask-Cors
openai>=1.0.0
python-dotenv
gunicorn
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
Werkzeug==2.2.3
PyJWT
